--- .env
@@ -1,2 +1,0 @@
-NEXT_PUBLIC_API_LOGIN=https://holaglowcontactsapidev.azurewebsites.net/Contact/Search
-API_NEWUSER=https://holaglowcontactsapidev.azurewebsites.net/Contact
