+++ .env.development
@@ -1,2 +1,1 @@
-NEXT_PUBLIC_API_LOGIN=https://holaglowcontactsapidev.azurewebsites.net/Contact/Search
-NEXT_PUBLIC_API_NEWUSER=https://holaglowcontactsapidev.azurewebsites.net/Contact
+NEXT_PUBLIC_CONTACTS_API=https://holaglowcontactsapidev.azurewebsites.net/
