+++ src/app/dashboard/SearchUser.tsx
@@ -9,7 +9,7 @@ const SearchUser: React.FC<SearchBarProps> = ({ email, handleFieldChange, handle
     await handleCheckUser();
     setIsLoading(false);
   };
-
+  
   return (
     <div id="buscar" className='w-3/4'>
       <div className="bg-gray-50 p-10 rounded-2xl flex">
@@ -40,5 +40,4 @@ const SearchUser: React.FC<SearchBarProps> = ({ email, handleFieldChange, handle
     </div>
   );
 };
-
 export default SearchUser;
